Defining movable bounds
First I tried to implement a flags system, but couldn't figure out  a simple, efficient way to check for the flag.
Then I tried to use Unity's built-in Tags system. But again, couldn't figure out how to obtain the Tag of an object at a relative position.
Finally I recognized that what I was trying to do was exactly the same as what we'd already done with the Hazard object; making an array of Vector3s, associating each Vector3 in the array with an object (which I still don't understand how that works under the hood, and that IS driving me nuts, but it's late and I need to move towards the goal, so if it works it works) so now instead of checking if the self-object's transform == the object in question, I'm adding a little mustard on the transform to check if the *future* position overlaps with walkable surface, and if it does, great, and if it doesn't, no movement will take place. Hopeully.