Can't figure out how to put a texture on the terrain.
I've read the documentation page about it, and the ones it links to, over and over, but where the instructions say to select a texture from a window that appears, the window I see contains only a few default options, and I can see no way to add any.
I give up for now, I'll just have to ask about it in class, which is in half an hour anyway.