Fixed the model
Got the spaceship to be the correct size, not really sure how, fugue state multiple tabs open, bits of advice from here and there, but it seems to have worked.
Also settled in on a good speed for the starfield to move at. Slower than in-game obstacles (like enemies which I don't have yet).
Lighting is bad and there's no sound or really any game mechanics, but it's more than nothing.