Deleted unnecessary assets
Also tried to make a cute milky way skybox but while Unity's documentation assures me I can adjust the horizon of the skybox, it is unwilling to explain where in the editor the options are to be found, god damn I hate the Unity documentation.
so I just wasted a huge amount of time on, once again, a little detail I thought would be simple to implement and would go a long way to making the game look nicer. But no of course we can't have that.